move functions to primitive_loader
