+++ alpha_automl/utils.py
@@ -419,48 +419,3 @@ from sklearn.pipeline import Pipeline""")
 \tpipeline.fit(X_train, y_train)
 \tprint(pipeline.predict(X_pred))
 """)
-
-import json
-from os.path import join, dirname
-from alpha_automl.primitive_loader import PRIMITIVE_TYPES
-
-PRIMITIVES_RANKING_PATH = join(dirname(__file__), 'resource/primitives_ranking.json')
-
-
-def load_primitives_ranking():
-    with open(PRIMITIVES_RANKING_PATH) as fin:
-        primitives = json.load(fin)
-    logger.debug('Ranking of all primitives loaded')
-
-    return primitives
-
-
-def record_primitive_performance(pipelines):
-    primitives_ranking = load_primitives_ranking()
-    
-    for idx, pipeline in enumerate(pipelines):
-        steps = pipeline.get_pipeline().steps
-        score = pipeline.get_score()
-        for step_name, _ in steps:
-            
-            if step_name not in PRIMITIVE_TYPES:
-                continue
-                
-            step_type = PRIMITIVE_TYPES[step_name]
-            if step_type not in primitives_ranking:
-                primitives_ranking[step_type] = {}
-            
-            if step_name not in primitives_ranking[step_type]:
-                primitives_ranking[step_type][step_name] = {"avg_score": 0, "runs": 0}
-            
-            ranking = primitives_ranking[step_type][step_name]
-            
-            weight = 1/(idx+1)
-            runs = ranking["runs"] + 1
-            avg_score = (ranking["avg_score"] * ranking["runs"] + score * weight) / runs
-            
-            primitives_ranking[step_type][step_name] = {"avg_score": avg_score, "runs": runs}
-    
-    with open(PRIMITIVES_RANKING_PATH, "w") as outfile:
-        json.dump(primitives_ranking, outfile)
-    return primitives_ranking
